Have the init script ignore hotplug interfaces that are already up. Closes: #838871
As pointed out by Enrico Weigelt, the sysvinit init script tries to
bring up hotplug interfaces. However, if udev is already trying to bring
that interface up, it will block the init script. This can delay the
boot process if the interface takes a long time to be brought up.

Use ifquery to check whether the interface is already (in the process of
being brought) up. If so, don't call ifup.