Fix race condition adding a static IPv6 default route on RA networks
As described in Bug#805445 there appears to be a race condition with
"inet6 static" configurations on networks where router advertisements
are also present.

The flow is like this

a - family inet is brought up
 a1 - ip link set dev %iface% up
 a2 - configure IPv4, execute hooks etc
b - family inet6 is brought up
 b1 - if gateway is set, set accept_ra 0
 b2 - ip link set dev %iface% up
 b3 - ip -6 route add default via %gateway%

However, starting at a1 the link is already up and the kernel might
have already learned an address and a default route from RA before the
sysctl variable is set. The address is flushed, but the default route
remains.

When in step b3 the default route is added it fails with

Aug 15 07:36:43 badwlrz-tljss1 ifup[575]: RTNETLINK answers: File exists
Aug 15 07:36:43 badwlrz-tljss1 ifup[575]: ifup: failed to bring up eth0

This ultimately leads to two things

1. networking.service fails, up stanzas from the interfaces are not executed
2. The RA learned default route still exists, but it not refreshed anymore
   (due to accept_ra=0). After the route is expired the system looses IPv6
   network connectivity.

Proposed fix is to use "route replace" instead of "route add". When the RA
learned route is still present it gets replaced, when it is not present the
route gets added.

Closes: #805445