Fix ping check loginc in ifupdown-wait-online.service. Closes: #952455
The ping command treats some network errors differently from others. In
particular, if there is an error response for an ICMP ECHO packet, it
exits immediately, while if there is no response and the ECHO is
silently dropped, it will continue sending packets. This makes it hard
to use to wait for a host to come online reliably. Work around this by
calling it in a loop until it succeeds, and use the timeout program to
enforce $WAIT_ONLINE_TIMEOUT.