Use relative names when executing programs
This switches program invocations to use relative names, so that we are
then not concerned about the filesystem layout of the system we are
installing into.

While for programs we ship and install we will know the end destination
and could simply replace that at build or install time, that seemed more
effort than simply making all usage relative.