Remove Conflicts=shutdown.target from ifup@.service
The Conflicts=shutdown.target directive has been added in this commit, as an alignment with networking.service:
https://salsa.debian.org/debian/ifupdown/-/commit/117cfa6

However, it breaks the intention of a previous commit, to not stop this unit on shutdown:
https://salsa.debian.org/debian/ifupdown/-/commit/3d06b08
Since networking.service/ifdown -a brings down all interfaces (including allow-hotplug) on shutdown already, having ifup@.service instances doing the same, potentially allows two ifdown calls to attempt to bring down the same interface concurrently. This has been found to cause shutdown hangups in rare cases.

In combination with DefaultDependencies=no (which overrides the default Conflicts=shutdown.target), Slice=system.slice (or using another appropriate slice) is needed for instantiated systemd units, to prevent them from being stopped on system shutdown:
https://www.freedesktop.org/software/systemd/man/latest/systemd.service.html#Default%20Dependencies
Re-adding Conflicts=shutdown.target (and Before=shutdown.target) inverts the key implication of DefaultDependencies=no, so that ifup@.service instances are again actively stopped on shutdown.

While Before=shutdown.target has no effect, when the service is not stopped at shutdown, it is removed here as well, as it was added as part of the same mistake.

Signed-off-by: MichaIng <micha@dietpi.com>